Performance fix
Now draw the grid and the lines in seperate layers that only gets
redrawn if the data changes.

Added orientation for the datapoint marker, so you can now select
datapoints horizontally as well now.

Also, shadows on the lines now stay after you update the lines. So, if
you trun on shadow, they stay on until you turn them off :)